@@ -1,62 +1,67 @@
 {
-  "name": "client",
-  "version": "0.1.0",
+  "name": "hack-party-client",
+  "version": "0.2.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
+    "lint": "eslint --ext .js,.vue --ignore-path .gitignore src",
+    "lint:fix": "eslint --ext .js,.vue --ignore-path .gitignore --fix src",
+    "test": "jest",
+    "test:inspect": "node --inspect-brk node_modules/.bin/jest --runInBand"
   },
   "engines": {
-    "node": "13.13.0",
-    "npm": "6.14.4"
+    "node": "17.6.0",
+    "npm": "8.5.1"
   },
   "dependencies": {
-    "@tailwindcss/ui": "^0.3.0",
-    "@vue/eslint-config-standard": "^5.1.2",
-    "axios": "^0.19.2",
-    "core-js": "^3.6.4",
-    "eslint-config-standard": "^14.1.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1",
-    "tailwindcss": "^1.4.4",
-    "twilio-sync": "^0.12.2",
-    "vue": "^2.6.11",
-    "vue-router": "^3.1.6",
-    "vue-slider-component": "^3.1.2",
-    "vuex": "^3.4.0",
-    "vuex-persistedstate": "^3.0.1"
+    "@heroicons/vue": "^1.0.5",
+    "@vueform/slider": "^2.0.9",
+    "axios": "^0.26.1",
+    "gsap": "^3.10.2",
+    "pinia": "^2.0.13",
+    "pinia-plugin-persistedstate": "^1.5.1",
+    "pug": "^3.0.2",
+    "pug-plain-loader": "^1.0.0",
+    "twilio-sync": "^3.1.0",
+    "vue": "^3.2.31",
+    "vue-router": "^4.0.13"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.3.0",
-    "@vue/cli-plugin-eslint": "~4.3.0",
-    "@vue/cli-plugin-router": "~4.3.0",
-    "@vue/cli-service": "~4.3.0",
-    "@vue/eslint-config-airbnb": "^5.0.2",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11"
+    "@babel/plugin-transform-runtime": "^7.17.10",
+    "@babel/preset-env": "^7.16.11",
+    "@pinia/testing": "^0.0.11",
+    "@vitejs/plugin-vue": "^2.3.0",
+    "@vue/eslint-config-standard": "^5.1.2",
+    "@vue/test-utils": "^2.0.0-rc.18",
+    "@vue/vue3-jest": "^27.0.0-alpha.4",
+    "autoprefixer": "^10.4.4",
+    "babel-plugin-transform-import-meta": "^2.1.1",
+    "babel-preset-vite": "^1.0.4",
+    "eslint": "^8.12.0",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-standard": "^5.0.0",
+    "eslint-plugin-vue": "^8.5.0",
+    "flush-promises": "^1.0.2",
+    "jest": "^27.5.1",
+    "lint-staged": "^12.3.7",
+    "postcss": "^8.4.12",
+    "regenerator-runtime": "^0.13.9",
+    "sass": "^1.49.9",
+    "tailwindcss": "^3.0.23",
+    "vite": "^2.9.0",
+    "vite-plugin-environment": "^1.1.1",
+    "yorkie": "^2.0.0"
   },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "standard"
-    ],
-    "parserOptions": {
-      "parser": "babel-eslint"
-    },
-    "rules": {}
+  "gitHooks": {
+    "pre-commit": "lint-staged"
   },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
+  "lint-staged": {
+    "*.{js,jsx,vue}": [
+      "npm run lint"
+    ]
+  }
 }
