@@ -1,5 +1,8 @@
 module.exports = {
   presets: [
-    '@vue/cli-plugin-babel/preset',
+    '@babel/preset-env'
   ],
-};
+  plugins: [
+    ['@babel/plugin-transform-runtime', { regenerator: true }]
+  ]
+}
