@@ -1,15 +1,13 @@
 module.exports = {
+  content: ['./index.html', './src/**/*.vue'],
   theme: {
     extend: {
       colors: {
-        green: "#00bc70",
-        blue: "#00b1e1",
-        orange: "#f36c3f",
-        darkgray: "#5b6670"
+        green: '#00bc70',
+        blue: '#00b1e1',
+        orange: '#f36c3f',
+        darkgray: '#5b6670'
       }
     }
-  },
-  plugins: [
-    require('@tailwindcss/ui'),
-  ]
+  }
 }
